Rename `RfDataRate.hpp` to `RfDataRateConfigs.hpp`
