+++ Sts1CobcSw/Rf/Rf.cpp
@@ -13,7 +13,7 @@
 #include <Sts1CobcSw/Hal/IoNames.hpp>
 #include <Sts1CobcSw/Hal/Spi.hpp>
 #include <Sts1CobcSw/Hal/Spis.hpp>
-#include <Sts1CobcSw/Rf/RfDataRate.hpp>
+#include <Sts1CobcSw/Rf/RfDataRateConfigs.hpp>
 #include <Sts1CobcSw/RodosTime/RodosTime.hpp>
 #include <Sts1CobcSw/Serial/Serial.hpp>
 #include <Sts1CobcSw/Utility/DebugPrint.hpp>  // IWYU pragma: keep
